requirements.txt
Flask==2.1.1
flask_bcrypt==1.0.1
flask_login==0.6.0
flask_migrate==3.1.0
flask_sqlalchemy==2.5.1
flask_wtf==1.0.1
python-dotenv==0.20.0
WTForms==3.0.1
